Rename spitfire.util to spitfire.text so there is only one "util" module
This change moves `spitfire.util` to `spitfire.text` since all of the
functionality in `spitfire.util` is text-related (e.g. unicode and i18n).  This
leaves `spitfire.compiler.util` as the only "util" module, which makes it clear
where imports are coming from.

Progress on #27
