+++ spitfire/text.py
